requirements.txt
python-dotenv
google-generativeai
Flask
flask-cors
NumPy
faiss-cpu
typing
